feat: json 파일에서 공지 데이터 가져오도록 개선 (#266)
